+++ src/Decision/PermissionDecisionFactoryInterface.php
@@ -1,0 +1,15 @@
+<?php
+
+declare(strict_types=1);
+
+namespace Sybio\PermissionBundle\Decision;
+
+use Symfony\Component\Validator\ConstraintViolationListInterface;
+
+interface PermissionDecisionFactoryInterface
+{
+    public function createDecision(
+        bool $granted,
+        ConstraintViolationListInterface $violations,
+    ): PermissionDecisionInterface;
+}